refactor(metrics): extract metrics logic into separate module
The metrics-related code has been modularized and moved to a new `metrics.py` file. This refactoring enhances code readability and separation of concerns by isolating the metrics functionality from the main application logic. The changes include:

- Moving Prometheus metrics counters and gauges to `metrics.py`.
- Refactoring functions related to logging configuration and metrics server startup into `metrics.py`.
- Adjusting imports and method calls in `app.py` to use the new `metrics` module.
- Updating tests to accommodate the new structure, ensuring metric-related operations are correctly tested.

This change aids maintainability and allows for easier future enhancements to metrics handling without affecting core application code.
